requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
Deprecated==1.2.13
fastapi==0.85.1
flower==1.2.0
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
humanize==4.4.0
idna==3.4
kombu==5.2.4
ndg-httpsclient==0.5.1
packaging==21.3
playwright==1.27.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyee==8.1.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1
